app/backend/requirements.txt
fastapi
uvicorn
redis
requests
openai
tiktoken
langchain
langchain_openai
python-dotenv
postgres
psycopg2-binary
psycopg[binary]==3.1.*
pgvector
langchain-community
langchain-postgres
unstructured
libmagic
nltk